Multinomial and Bernoulli Naïve Bayes classifiers
This project focuses on text classification, a popular machine learning problem. The goal is to build a classifier to identify the subreddit from which a comment originated accurately. Several algorithms were used in this project, including Multinomial and Bernoulli Naïve Bayes classifiers, Support Vector Machines, and Stochastic Gradient Descent classifiers. The raw text comments needed to be converted into features to achieve the best performance from these classifiers. The majority of this report explains the methods used to construct the features that produced the best results. Using 5-fold cross-validation, the Stochastic Gradient Descent classifier with logistic regression yielded the best results.